functions/analyze_sentiment_function/requirements.txt
functions-framework==3.*
google-cloud-pubsub
google-cloud-language>=2.4.1